slight UI adjustment
