+++ src/pages/about.tsx
@@ -3,8 +3,7 @@ import React from 'react'
 
 const AboutPage = () => {
     return (
-        <main className='min-h-screen'>
-            <Countdown date={new Date("2023-09-07T06:39:00+0700")}/>
+        <main className='min-h-screen bg-black-1'>
         </main>
     )
 }
